fix(core): robustness pass — validation, leak-safety, Int64 counts, real metrics
Hardening from a Codex-reviewed audit (Critical/High/Medium):

- Validate config in the constructor: non-null serializer/comparer,
  MergeWayCount>=2 (a 1-way merge never converges → was an infinite
  loop), MaxMemoryBytes/BufferSize>0, DegreeOfParallelism>=1,
  non-empty TempDirectory, EstimatedItemSize>0. Sort() null-guards
  its streams.
- ComputeChunkCapacity(): long-space division clamped to
  [1, int.MaxValue] so a huge budget / tiny item can't overflow the
  int cast or produce a zero-size buffer.
- Resource safety: all input BinaryReaders use leaveOpen:true (the
  caller owns the stream) and are disposed; MergeBatch builds its
  reader list inside the try and disposes in finally; RS closes a
  half-written chunk in finally on fault.
- Parallel pipeline: a reader-thread fault now cancels + drains the
  workers before propagating, so no worker writes into tempDir after
  Sort() starts deleting it; a worker fault is no longer masked by the
  reader's induced OperationCanceledException; single inner exceptions
  rethrow via ExceptionDispatchInfo (stack preserved).
- Int64 count headers everywhere (chunk files + final output) so sorts
  beyond int.MaxValue items don't corrupt counts. Console verifier and
  test reader updated.
- SortMetrics.BytesRead/BytesWritten now populated (were dead fields):
  cheap Position delta for seekable streams, CountingStream wrapper
  fallback for non-seekable — no hot-path cost on the common case.
- Document stream ownership + unstable-sort semantics on IExternalSorter,
  and the binary record format on RecordSerializer.

Tests: +19 in RobustnessTests (validation, null guards, fault cleanup
on all three chunk strategies, a cross-platform proof that the parallel
fault path drains blocked workers, Int64 header, truncated-input EOF,
leaveOpen, BytesRead/Written). Suite 51 -> 70, all green. Console e2e
re-verified (200k records / 2 MB RAM).

+++ src/ExternalSorting.Console/Program.cs
@@ -101,10 +101,10 @@ bool isSorted = true;
 using (var fs = File.OpenRead(outputPath))
 using (var br = new BinaryReader(fs))
 {
-    int itemCount = br.ReadInt32();
+    long itemCount = br.ReadInt64();
     SortRecord? prev = null;
 
-    for (int i = 0; i < itemCount; i++)
+    for (long i = 0; i < itemCount; i++)
     {
         var current = serializer.Read(br);
         if (prev.HasValue && current.CompareTo(prev.Value) < 0)
